+++ gcs/CHANGES.md
@@ -1,5 +1,8 @@
 ### 2.2.9 - 2022-XX-XX
 
+1.  The Google Cloud Storage Connector now can be used as a
+    [Hadoop Credential Provider](https://hadoop.apache.org/docs/current/hadoop-project-dist/hadoop-common/CredentialProviderAPI.html).
+
 ### 2.2.8 - 2022-09-12
 
 1.  Set socket read timeout (`fs.gs.http.read-timeout`) as early as possible on